UART: Change default baudrate to 19200
Previously serial port was using 115200, i.e. 12000000/6/16,
but that gives actually 125000. Round down - 5/16 - gives
150000, so too far from the 115200 target. Rouding up give
- 7/16 - gives 107142.9, better, but still far from the target.
Getting down to a lower baudrate 19200 reduces the error
on that TX implementation (we need the division by 16) so
we can use the same 'uart_clock' to sample as well in the TX
FSM right in the middle of the asserted bit.