UART: Reduce time constraint on synthesis
For some nasty reason, probably related to the current UART
implementation, possibly inside fifo.v code, the time
constraint on synthesis got even tighter, so we se to 3 MHz
just to satisfy the compiler. TODO: understand what's the
hindrance in the code to reach out higher speed on synthesis,
like 400 or so MHz.