UART: Add first RX version that really works!
RX part now works, although it doesn't use a FIFO and instead
only uses one single register to keep the received data. It
seems the FIFO is broken somehow to be used by the RX part and
so needs a refactor too before it can be used again by the RX.
That said, *finally* the RX waveform sync mechanism is working
and its clock implementation is trivial. Default baudrate set
is 9600.